+++ src/main/java/com/mapcode/DataModel.java
@@ -19,6 +19,7 @@ package com.mapcode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.annotation.Nonnull;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
@@ -30,14 +31,8 @@ import java.io.InputStream;
  *
  * This class contains the module that reads the Mapcode areas into memory and processes them.
  */
-class DataAccess {
-    private static final Logger LOG = LoggerFactory.getLogger(DataAccess.class);
-
-    private static final int NR_TERRITORIES;
-    private static final int NR_TERRITORY_RECORDS;
-
-    private static final int[] INDEX;
-    private static final int[] DATA;
+class DataModel {
+    private static final Logger LOG = LoggerFactory.getLogger(DataModel.class);
 
     private static final int HEADER_ID_1 = 0;
     private static final int HEADER_ID_2 = 1;
@@ -65,18 +60,49 @@ class DataAccess {
     private static final int POS_INDEX_FIRST_RECORD = 0;
     private static final int POS_INDEX_LAST_RECORD = 1;
 
-    private static final String FILE_NAME = "/com/mapcode/mminfo.dat";
+    private static final String DATA_FILE_NAME = "/com/mapcode/mminfo.dat";
     private static final int FILE_BUFFER_SIZE = 50000;
 
     private static final int DATA_VERSION_MIN = 220;
 
-    // Read data only once in static initializer.
-    static {
-        LOG.info("DataAccess: reading regions from file: {}", FILE_NAME);
+    private static int readIntLoHi(final int lo, final int hi) {
+        return (lo & 0xff) + ((hi & 0xff) << 8);
+    }
+
+    private static int readLongLoHi(final int lo, final int mid1, final int mid2, final int hi) {
+        return ((lo & 0xff)) + ((mid1 & 0xff) << 8) + ((mid2 & 0xff) << 16) + ((hi & 0xff) << 24);
+    }
+
+    // Data.
+    private final int nrTerritories;
+    private final int nrTerritoryRecords;
+
+    private final int[] index;
+    private final int[] data;
+
+    @SuppressWarnings("StaticNonFinalField")
+    private static volatile DataModel instance = null;
+    private static final Object mutex = new Object();
+
+    @SuppressWarnings({"DoubleCheckedLocking", "SynchronizationOnStaticField"})
+    public static DataModel getInstance() {
+        if (instance == null) {
+            synchronized (mutex) {
+                if (instance == null) {
+                    instance = new DataModel(DATA_FILE_NAME);
+                }
+            }
+        }
+        return instance;
+    }
+
+    DataModel(@Nonnull final String fileName) throws IncorrectDataModelException {
+        // Read data only once in static initializer.
+        LOG.info("DataAccess: reading regions from file: {}", fileName);
         final byte[] readBuffer = new byte[FILE_BUFFER_SIZE];
         int total = 0;
         try {
-            final InputStream inputStream = DataAccess.class.getResourceAsStream(FILE_NAME);
+            final InputStream inputStream = DataModel.class.getResourceAsStream(fileName);
             try {
                 final ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
                 try {
@@ -94,49 +120,53 @@ class DataAccess {
                     assert total == bytes.length;
                     if (total < 12) {
                         LOG.error("DataAccess: expected more than {} bytes", total);
-                        throw new IllegalStateException("Data file corrupt: " + FILE_NAME);
+                        throw new IncorrectDataModelException("Data file corrupt: " + fileName);
                     }
 
                     // Read "MC", VERSION.
                     assert total > 8;  // "MC" (2) + VERSION (2) + NR TERRITORIES (2) + NR TERRITORY RECORDS (2).
-                    assert (char) bytes[HEADER_ID_1] == 'M';
-                    assert (char) bytes[HEADER_ID_2] == 'C';
+                    if ((bytes[HEADER_ID_1] != 'M') || (bytes[HEADER_ID_2] != 'C')) {
+                        throw new IncorrectDataModelException("Data file does not start with correct header: " + fileName);
+                    }
                     final int dataVersion = readIntLoHi(bytes[HEADER_VERSION_LO], bytes[HEADER_VERSION_HI]);
-                    assert dataVersion >= DATA_VERSION_MIN;
+
+                    if (dataVersion < DATA_VERSION_MIN) {
+                        throw new IncorrectDataModelException("Data file version " + dataVersion + " too low: " + fileName);
+                    }
 
                     // Read header: NR TERRITORIES, NR RECTANGLE RECORDS.
-                    NR_TERRITORY_RECORDS = readIntLoHi(bytes[HEADER_NR_TERRITORIES_RECS_LO], bytes[HEADER_NR_TERRITORIES_RECS_HI]);
-                    NR_TERRITORIES = readIntLoHi(bytes[HEADER_NR_TERRITORIES_LO], bytes[HEADER_NR_TERRITORIES_HI]);
+                    nrTerritoryRecords = readIntLoHi(bytes[HEADER_NR_TERRITORIES_RECS_LO], bytes[HEADER_NR_TERRITORIES_RECS_HI]);
+                    nrTerritories = readIntLoHi(bytes[HEADER_NR_TERRITORIES_LO], bytes[HEADER_NR_TERRITORIES_HI]);
 
                     // Check if the number of territories matches the enumeration in Territory.
-                    if (NR_TERRITORIES != Territory.values().length) {
-                        LOG.error("DataAccess: expected {} territories, got {}", Territory.values().length, NR_TERRITORIES);
-                        throw new IllegalStateException("Data file corrupt: " + FILE_NAME);
+                    if (nrTerritories != Territory.values().length) {
+                        LOG.error("DataAccess: expected {} territories, got {}", Territory.values().length, nrTerritories);
+                        throw new IncorrectDataModelException("Data file corrupt: " + fileName);
                     }
 
                     // Check if the expected file size matched what we found.
                     final int expectedSize = HEADER_SIZE +
-                            ((NR_TERRITORIES + 1) * BYTES_PER_INT) +
-                            (NR_TERRITORY_RECORDS * (DATA_FIELDS_PER_REC * BYTES_PER_LONG));
+                            ((nrTerritories + 1) * BYTES_PER_INT) +
+                            (nrTerritoryRecords * (DATA_FIELDS_PER_REC * BYTES_PER_LONG));
 
                     if (expectedSize != total) {
                         LOG.error("DataAccess: expected {} bytes, got {}", expectedSize, total);
-                        throw new IllegalStateException("Data file corrupt: " + FILE_NAME);
+                        throw new IncorrectDataModelException("Data file corrupt: " + fileName);
                     }
-                    LOG.debug("DataAccess: version={} territories={} territory records={}", dataVersion, NR_TERRITORIES, NR_TERRITORY_RECORDS);
+                    LOG.debug("DataAccess: version={} territories={} territory records={}", dataVersion, nrTerritories, nrTerritoryRecords);
 
                     // Read DATA+START array (2 bytes per territory, plus closing record).
-                    INDEX = new int[NR_TERRITORIES + 1];
+                    index = new int[nrTerritories + 1];
                     int i = HEADER_SIZE;
-                    for (int k = 0; k <= NR_TERRITORIES; k++) {
-                        INDEX[k] = readIntLoHi(bytes[i], bytes[i + 1]);
+                    for (int k = 0; k <= nrTerritories; k++) {
+                        index[k] = readIntLoHi(bytes[i], bytes[i + 1]);
                         i += 2;
                     }
 
                     // Read territory rectangle data (DATA_FIELDS_PER_REC longs per record).
-                    DATA = new int[NR_TERRITORY_RECORDS * DATA_FIELDS_PER_REC];
-                    for (int k = 0; k < (NR_TERRITORY_RECORDS * DATA_FIELDS_PER_REC); k++) {
-                        DATA[k] = readLongLoHi(bytes[i], bytes[i + 1], bytes[i + 2], bytes[i + 3]);
+                    data = new int[nrTerritoryRecords * DATA_FIELDS_PER_REC];
+                    for (int k = 0; k < (nrTerritoryRecords * DATA_FIELDS_PER_REC); k++) {
+                        data[k] = readLongLoHi(bytes[i], bytes[i + 1], bytes[i + 2], bytes[i + 3]);
                         i += 4;
                     }
                 } finally {
@@ -148,22 +178,10 @@ class DataAccess {
                 inputStream.close();
             }
         } catch (final IOException e) {
-            throw new ExceptionInInitializerError("Cannot initialize static data structure from: " +
-                    FILE_NAME + ", exception=" + e);
+            throw new IncorrectDataModelException("Cannot initialize static data structure from: " +
+                    fileName + ", exception=" + e);
         }
         LOG.info("DataAccess: regions initialized, read {} bytes", total);
-    }
-
-    private static int readIntLoHi(final int lo, final int hi) {
-        return (lo & 0xff) + ((hi & 0xff) << 8);
-    }
-
-    private static int readLongLoHi(final int lo, final int mid1, final int mid2, final int hi) {
-        return ((lo & 0xff)) + ((mid1 & 0xff) << 8) + ((mid2 & 0xff) << 16) + ((hi & 0xff) << 24);
-    }
-
-    private DataAccess() {
-        // Empty.
     }
 
     /**
@@ -171,8 +189,8 @@ class DataAccess {
      *
      * @return Number of territories.
      */
-    static int getNrTerritories() {
-        return NR_TERRITORIES;
+    int getNrTerritories() {
+        return nrTerritories;
     }
 
     /**
@@ -180,42 +198,42 @@ class DataAccess {
      *
      * @return Number of rectangles per territory.
      */
-    static int getNrTerritoryRecords() {
-        return NR_TERRITORY_RECORDS;
+    int getNrTerritoryRecords() {
+        return nrTerritoryRecords;
     }
 
     @SuppressWarnings("PointlessArithmeticExpression")
-    static int getLonMicroDegMin(final int territoryRecord) {
-        return DATA[((territoryRecord * DATA_FIELDS_PER_REC) + POS_DATA_LON_MICRO_DEG_MIN)];
+    int getLonMicroDegMin(final int territoryRecord) {
+        return data[((territoryRecord * DATA_FIELDS_PER_REC) + POS_DATA_LON_MICRO_DEG_MIN)];
     }
 
-    static int getLatMicroDegMin(final int territoryRecord) {
-        return DATA[(territoryRecord * DATA_FIELDS_PER_REC) + POS_DATA_LAT_MICRO_DEG_MIN];
+    int getLatMicroDegMin(final int territoryRecord) {
+        return data[(territoryRecord * DATA_FIELDS_PER_REC) + POS_DATA_LAT_MICRO_DEG_MIN];
     }
 
-    static int getLonMicroDegMax(final int territoryRecord) {
-        return DATA[(territoryRecord * DATA_FIELDS_PER_REC) + POS_DATA_LON_MICRO_DEG_MAX];
+    int getLonMicroDegMax(final int territoryRecord) {
+        return data[(territoryRecord * DATA_FIELDS_PER_REC) + POS_DATA_LON_MICRO_DEG_MAX];
     }
 
-    static int getLatMicroDegMax(final int territoryRecord) {
-        return DATA[(territoryRecord * DATA_FIELDS_PER_REC) + POS_DATA_LAT_MICRO_DEG_MAX];
+    int getLatMicroDegMax(final int territoryRecord) {
+        return data[(territoryRecord * DATA_FIELDS_PER_REC) + POS_DATA_LAT_MICRO_DEG_MAX];
     }
 
-    static int getDataFlags(final int territoryRecord) {
-        return DATA[(territoryRecord * DATA_FIELDS_PER_REC) + POS_DATA_DATA_FLAGS] & MASK_DATA_DATA_FLAGS;
+    int getDataFlags(final int territoryRecord) {
+        return data[(territoryRecord * DATA_FIELDS_PER_REC) + POS_DATA_DATA_FLAGS] & MASK_DATA_DATA_FLAGS;
     }
 
-    static int getSmartDiv(final int territoryRecord) {
-        return DATA[(territoryRecord * DATA_FIELDS_PER_REC) + POS_DATA_DATA_FLAGS] >> SHIFT_POS_DATA_SMART_DIV;
+    int getSmartDiv(final int territoryRecord) {
+        return data[(territoryRecord * DATA_FIELDS_PER_REC) + POS_DATA_DATA_FLAGS] >> SHIFT_POS_DATA_SMART_DIV;
     }
 
     // Low-level routines for data access.
     @SuppressWarnings("PointlessArithmeticExpression")
-    static int getDataFirstRecord(final int territoryNumber) {
-        return INDEX[territoryNumber + POS_INDEX_FIRST_RECORD];
+    int getDataFirstRecord(final int territoryNumber) {
+        return index[territoryNumber + POS_INDEX_FIRST_RECORD];
     }
 
-    static int getDataLastRecord(final int territoryNumber) {
-        return INDEX[territoryNumber + POS_INDEX_LAST_RECORD] - 1;
+    int getDataLastRecord(final int territoryNumber) {
+        return index[territoryNumber + POS_INDEX_LAST_RECORD] - 1;
     }
 }